LayoutTests/inspector/debugger/scripts-file-selector.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>

<script>
function test()
{
    // Always use indentation in scripts drop box
    WebInspector._isMac = false;

    WebInspector.showPanel("scripts");

    var navigator = new WebInspector.ScriptsNavigator();
    var workspace = new WebInspector.Workspace();
    var workspaceProvider = new WebInspector.SimpleWorkspaceProvider(workspace, WebInspector.projectTypes.Network);

    var uiSourceCodes = [];
    function addUISourceCode(url, isContentScript)
    {
        var contentProvider = new WebInspector.StaticContentProvider(WebInspector.resourceTypes.Script, "", "text/javascript");
        var uiSourceCode = workspaceProvider.addFileForURL(url, contentProvider, false, isContentScript);
        navigator.addUISourceCode(uiSourceCode);
        uiSourceCodes.push(uiSourceCode);
    }

    var rootURL = "http://localhost:8080/LayoutTests/inspector/debugger/";
    addUISourceCode(rootURL + "foo/bar/script.js", false);
    addUISourceCode(rootURL + "foo/bar/contentScript2.js?a=1", true);
    addUISourceCode(rootURL + "foo/bar/script.js?a=2", false);
    addUISourceCode(rootURL + "foo/bar/contentScript.js?a=2", true);
    addUISourceCode(rootURL + "foo/bar/script.js?a=1", false);
    addUISourceCode(rootURL + "foo/baz/script.js", false);
    addUISourceCode(rootURL + "foo/bar/contentScript.js?a=1", true);
    addUISourceCode("http://example.com/?a=b", false);
    addUISourceCode("?a=b", false);
    addUISourceCode("very_looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong_url", false);

    InspectorTest.dumpScriptsNavigator(navigator);
    for (var i = 0; i < uiSourceCodes.length; ++i)
        navigator.revealUISourceCode(uiSourceCodes[i]);
    InspectorTest.dumpScriptsNavigator(navigator);

    InspectorTest.completeTest();
}
</script>

</head>
<body onload="runTest()">
<p>
Tests scripts panel file selectors.
</p>
</body>

</html>
