LayoutTests/inspector/debugger/scripts-file-selector-expected.txt

Tests scripts panel file selectors.

Dumping ComboBoxFileSelector:
    ?a=b(?a=b)
    very_looooooooo...oooooooong_url(very_looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong_url)
— example.com —
    ?a=b(http://example.com/?a=b)
— localhost —
LayoutTests/inspector/debugger/foo/bar
    script.js(<root>/foo/bar/script.js)
    script.js?a=1(<root>/foo/bar/script.js?a=1)
    script.js?a=2(<root>/foo/bar/script.js?a=2)
LayoutTests/inspector/debugger/foo/baz
    script.js(<root>/foo/baz/script.js)
— Content scripts —
LayoutTests/inspector/debugger/foo/bar
    contentScript.js?a=1(<root>/foo/bar/contentScript.js?a=1)
    contentScript.js?a=2(<root>/foo/bar/contentScript.js?a=2)
    contentScript2.js?a=1(<root>/foo/bar/contentScript2.js?a=1)
Dumping ScriptsNavigator 'Scripts' tab:
  example.com
    ?a=b
  localhost
    LayoutTests/inspector/debugger/foo/bar
      script.js
      script.js?a=1
      script.js?a=2
    LayoutTests/inspector/debugger/foo/baz
      script.js
  ?a=b
  very_looooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong_url
Dumping ScriptsNavigator 'Content scripts' tab:
  localhost
    LayoutTests/inspector/debugger/foo/bar
      contentScript.js?a=1
      contentScript.js?a=2
      contentScript2.js?a=1
